Show a more detailed scan progression
Until now, api_scan used a zap.ascan.status() call to get the completion
percent (every 10 seconds)
This causes problems in massive API, where it can remain at 0% for
substential time, resulting in the user not knowing if the scan is stuck
on ongoing.

This is an attempt to improve that situation, by also sending a
zap.ascan.scan_progress() request.

This patch propose to cumulate all the "message" sent by all plugins,
which should be an evergrowing number showing that there is an
evolution (but we also keep the regular %age of completion)

NOTE: I think the "message" corresponds to a message sent from the
plugin to ZAP (not the plugin to the target), which would explain that
this number grows slowly

Example of return :

[
  "https://console.stage.redhat.com",
  {
    "HostProcess": [
      {
        "Plugin": [
          "Remote File Inclusion",  # plugin name
          "7",                      # plugin ID
          "release",                # version (release or beta)
          "7%",                     # %completion, Pending or Complete
          "1002868",                # time spent
          "4390",                   # number of "messages"
          "0"                       # number of alerts found
        ]
      },
      {
        "Plugin": [
          "External Redirect",
          "20019",
          "release",
          "Pending",
          "0",
          "0",
          "0"
        ]
      },
...
    ]
  }
]