Introducing a new podman wrapper (#46)
* Introducing a new podman wrapper

The main reason for the wrapper is to avoid the need to run `unshare`,
by formatting a defined user mapping, such that the `zap` user maps the
host user. This way ZAP can write in the `./results` share without
needing to change ownership.

source:
https://github.com/containers/podman/blob/main/troubleshooting.md#39-podman-run-fails-with-error-unrecognized-namespace-mode-keep-iduid1000gid1000-passed
In very recent podman version, this hack can be simplified.

This change is backward compatible : nothing prevents the user to use
any old methods for starting RapiDAST.

Note:
Unlike `runenv.sh`, this script does *not* attempt to stop the container
in case it is already running. The reasoning is : if RapiDAST is
currently undergoing a looooong scan, a user probably prefers the new
command to fail, rather than cancelling the current scan.
However, we could improve that.

Other minor changes:
In README.md:
- removed the reference to `[podman|docker] pull`, as it refers to a
  different image, and afaik, will be done by the compose command
- updated with the new command, removed references to unshare
- few minor readability/consistency update

In runenv.sh / runenv-ui.sh: echo "deprecated" message

* MacOS compatibility in podman-wrapper.sh

Co-authored-by: cedric <cedric@ltop.lan>