Introducing a new RapiDAST
This commit proposes a brand new RapiDAST.

Major change is that it will be easier to add additional scanners aside
of ZAP.
ZAP also now works in "Automation" mode instead of using the ZAP API.

The overall design is more user friendly:
- The configuration file has been simplified, keeping only the necessary
  entries
- The configuration schema is versionned: when RapiDAST encounters an
  older schema version, it will attempt to update it, such that users
  will not be forced to always keep up with the changes.

To use RapiDAST:
1) Generate a config file (usee config/config-template.yaml)
2) Run `./rapidast.py --config <config.yaml>`
3) Results will be stored in the `results/` directory

For the internal changes:
- RapiDAST will spawn scanners (currentl only ZAP) in a container
  (currently only podman)
- The ZAP scanner will generate an Automation config based on the
  RapiDAST configuration, and send it to the ZAP image.
- RapiDAST will a best effort to remain compatible with configuration
  generated for the old RapiDAST, but the compatibility is not 100%

For further information about usage and development, see the following
files:
+ docs/DEVELOPER-GUIDE.md
+ README.md (the user guide)
+ config/config-template.yaml  (example configuration)
+ config/config-template-long.yaml  (complete configuration)

A currently missing feature, compared to previous version, is the
operator mode. This may be re-added in future version.

Cookie authentication is also currently missing, but will be re-added
soon.