[bug][configmodel] crash when an entry is None
`config.exists()` did not work correctly when an entry was set to None.
This is because getting `None["x"]` returns a `TypeError` instead of the
usual `KeyError`, and thus the exception was not caught, crashing
RapiDAST.

This fixes the issue by catching this particular exception.

The issue was made visible when converting from older schema when
`general` existed but was empty.

eg.:

```yaml
config:
        configVersion: 1
general:
```

Also added a test dedicated to the `config.exists()` function.