[ZAP] workaround for OAuth2 download of OpenAPI
This is aimed at being temporary only.
If ZAP enables authenticated OpenAPI download, it can be removed.
Alternatively, the code could be improved slowly in the future.

-> added a `scanners.zap.miscOptions.oauth2OpenapiManualDownload` config
entry. It defaults to `False`
-> if the above is True, *and* authentication is OAuth2 *and* there is a
OpenAPI definition downloaded over URL, then the code downloads it, and
modifies `apiUrl` to `apiFile`
-> a new `scanners.downloaders.authenticated_download_with_rtoken()`
function is provided for this matter.

Note:
Because of the nature of the action, I don't see how to make a pytest
for this.