[ZAP] re-added XML as a report format
Defect Dojo uses ZAP's XML format for import. It also accepts SARIF, but
using ZAP's format might be prefered.

Also changed :
- simplified the format selection, to avoid if/else list
- removed the globbing: all the reports are in a dedicated directory in
  the container, this way it's easier to extract them (including html's
  subdir)
- scanner.zap.report.format now accepts either a string or a list[*]

The behaviour has not changed:
- the default (either no format, or all formats were invalid) is to have
  a single `json` format
- pytest remain functional as previously

[*]`format` is singular (no 's'), so it might have lead to confusion
with regards to the content needing to be a list.
If it's a string, it will be converted to a list of a single element.