[experimental] support Defect Dojo
This is a first commit that shows a Proof of Concept for Defect Dojo
integration.

The idea is that RapiDAST adds an optional 6th step for each scanner.

This 6th step is, if the scanner has the capability, to provide the scan
information for Defect Dojo.
Rapidast, then push this to Defect Dojo using URL & authentication
provided via the configuration file.

Additional notes:
* In ZAP, when DD is enabled, we enforce the XML report (in addition to
  whichever report(s) may have been set).
  In the code, the report list was changed to a set, so that we do not
  have to bother about duplicate reports

* templates & READMEs have been updated with docs for both user and devs

* currently, there is a lack of test functions. That will arrive in the
  next iteration.

Path forward and improvement:
Ideally, the parent `RapidastScanner` should be able to make more heavy
lifting, so that the scanners themselves don't need to duplicate too
much code.
In other word: `data_for_defect_dojo()` should move to parent, and the
scanner themselves should be able to only provide a filename and default
values.
That can be worked out when we are okay with the process.