[IMG] Build a multiuser image
Creates an image with ZAP and RapiDAST in /opt, accessible from any user
Does not enforce specific user

Run `run_rapidast.sh <config>` to start rapidast with a config file

Note:
Because the shell script changes the current working directory, it is
strongly advised to set `config.base_results_dir` to an absolute
directory, writable by the user (e.g.: `/tmp/results/`). If not set, the
default location will go to /opt/rapidast/results/, which may not be
writable.