[rapidast] some error handling
Catching some errors and prints a better explanation on the console when
it happens.

In particular :
- when podman is not in the PATH
- when a scanner is not found, or the type not implemented

Also fixing emerging pylint errors (too many statements and branching in
the `run()` function), by creating a new function: `run_scanner()` which
runs only 1 scanner. This way, there are 2 smaller functions instead of
a giant one.

Maybe we could also implement pytests for this new `run_scanner()`