[ZAP-flatpak] Undo creation of separate home dir
A previous commit was aimed at creating a new ZAP home directory, to
force the flatpak instance to have a fresh environment every time.

It does not work well, and ZAP crashes with the following error:
The mandatory add-on was not found: network

This fix undo it, and the classic `~/.ZAP/` is being used again.