[ZAP]regression workaround for MacOS missing addon
This is a workaround for upstream ZAP issue 7703
https://github.com/zaproxy/zaproxy/issues/7703

Waiting for the upstream fix to be pulled in a stable release.

Issue description:
ZAP attempts to delete outdated plugins. The problem is that, in MacOS,
the installation is done as the user, and thus the user has permission
to delete the installation's plugins, and most of the plugins are
deleted.
The updated version of those plugins are stored in the user's home ZAP
directory, but RapiDAST no longer use this directory.

The result is: ZAP refuses to start because 2 mandatory plugins are
missing.

The workaround:
When the mandatory plugins are found to be missing, this happens:

1) the plugins are manually downloaded, straight from upstream.
This is sufficient to run ZAP, but not sufficient to run RapiDAST,
because other plugins are still missing (e.g.: automation plugin)

2) RapiDAST then installs *all* plugins.

The downside of this is a longer runtime, because not only RapiDAST must
be run 1 more time, it also has to download all plugins.
The other downside is size: the plugin are roughly ½GB in size. This
will be deleted after the RapiDAST run (in case of success).

To avoid the downside, in MacOS, you can do the following:
1- Reinstall ZAP, but DO NOT RUN IT! (or it will delete the installation
plugins)
2- change the ownership of the ZAP application to `root`, so that when
run as a regular user, the deletion of the plugins will fail.
3- then you can run ZAP and RapiDAST normally