[ZAP] Allow user to override default Java max heap
zap.sh calculates the total RAM, divides by 4 and use that for the
`-Xmx` java option.
Issue: When RapiDAST runs on a dedicated environment (e.g.: pod, VM,
etc.), most of the ¾ RAM remaining is wasted.
This commits allows to override the value, which is necessary for
parsing large OpenAPI files (several MBs)

pytest added