[containerize] Create optional default config
If there is a file called 'rapidast-defaults.yaml' located in the same
directory as `rapidast.py`, then this file is loaded as a YAML file.
It content will be merged to the current rapidast config BUT without
overwriting any existing entries.

In practice, this allows anyone to create default values.

Containerfile uses that to set `general.container.type` to 'none'
This way, if a config file does not set a container type for a given
scanner, `none` will be chosen.