[templates] container type set to default
Commenting out the container.type, so that templates rely on default.
Except in `generic` scanner, where `podman` is the only supported type,
so it makes sense to enforce it at the moment.

Relying on default will allow RapiDAST to use podman on a host, and none
in the RapiDAST image.

Additional changes: I tried to normalize documentation within the YAML
templates.
e.g.:

```
```

Is that a entry called `Optional` commented out or is it documentation?

Now:
- if there is no space between the `#` and the first letter, it's an
  entry commented out
- if there is a space between `#` and the first letter, and that letter
  is upper case, it's probably documentation