[podman_wrapper] simplify gather UID/GID mapping
Originally, the method was a strict rewrite from
https://github.com/containers/podman/blob/main/troubleshooting.md#39-podman-run-fails-with-error-unrecognized-namespace-mode-keep-iduid1000gid1000-passed
Which contained a terrible eval() call

We modified it in commit 3c7e1c8f, creating a "safe_eval".

However, this does not work well on MacOS

Rewriting the methodology from scratch: retreive the JSON output, and
make the calculations from there.

Added pytest for it