[ZAP] oauth2 pre-authentication mode
The OAuth2/refresh-token authentication sometimes behaves unexpectedly.

This commit adds a `authenticated.parameters.preauth` optional
parameter.
When set to `true`, it will generate a single Bearer token, and use that
token throughout the scan.
The downside is that the token will never be refreshed, and thus once
expired, the target will (should) no longer consider it valid

Also added pytest, but with a limited testing: we just verify that the
plugin attempts to get a token a fails