[ZAP] accepts report.format as string
Previously, when `report.format` was provided as a string instead of a
list, it would be forced into a list (i.e.: "xml" becomes the list ["x",
"m", "l"].

Because `format` is singular, it might be expected that a user provides
a string instead of a list with a single element.
If this happens, the string is now turned into a list of a single
element.

Also: update pre_commit's pylint version.
The previous version was no longer compatible with python version 3.12.
Resulting in pre_commit crash.