ZAP: change HOME if it is not writable
Firefox requires a home directory.
When crawling (Ajax spider), in Openshift, Firefox is unable to start if
it can't write a ~/.firefox

This is an issue in Openshift, where the user is created on the fly and
its home directory is '/'.

In that case, create a temporary directory, and assign HOME to it.