podman wrapper: uses userns option when possible
On podman >= 4.3, the --userns keep-id:uid=X,gid=Y option allows to map
the current user to a specific container user.

Previous code: we created an entire UID/GID mapping to make this happen

In this change: if podman >= 4.3, we simply use the userns option,
otherwise we fallback to the previous workaround.

Also:
- in the workaround: only do the mapping when subUID/subGID are configured.
- Comments improved (hopefully)
- Updated README troubleshooting