feat: add cookie authentication e2e test and refactor to parameterized tests (#413)
* feat: add cookie authentication e2e test and refactor to parameterized tests

Add cookie authentication e2e test and refactor all authentication tests
to use pytest parameterization

* refactor(e2e-tests): remove unnecessary executable from ZAP config

The 'executable: zap.sh' parameter is the default for the ZAP scanner
and does not need to be explicitly defined in the configuration.

This commit removes the redundant container parameter from the
E2E authentication test manifests to simplify the configuration.